forge fmt

+++ src/POC.sol
@@ -1,13 +1,10 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.13;
 
-
 contract POC {
-
-    uint public pocCheck = 2;
+    uint256 public pocCheck = 2;
 
     address owner = 0x4838B106FCe9647Bdf1E7877BF73cE8B0BAD5f97;
 
     string hello = "scroll";
-
 }
